refactor(Rendering): extract ScalarColoring into a helper
