+++ Sources/Rendering/Core/Mapper2D/index.d.ts
@@ -1,98 +1,32 @@
-import { Nullable } from '../../../types';
 import vtkAbstractMapper, {
   IAbstractMapperInitialValues,
 } from '../AbstractMapper';
+import { IPrimitiveCount } from '../Mapper';
+import {
+  IScalarColoringInitialValues,
+  TScalarColoringWithout,
+} from '../Mapper/ScalarColoringHelper';
 
-export enum ColorMode {
-  DEFAULT,
-  MAP_SCALARS,
-  DIRECT_SCALARS,
-}
+export { ColorMode, GetArray, ScalarMode } from '../Mapper/Constants';
 
-export enum ScalarMode {
-  DEFAULT,
-  USE_POINT_DATA,
-  USE_CELL_DATA,
-  USE_POINT_FIELD_DATA,
-  USE_CELL_FIELD_DATA,
-  USE_FIELD_DATA,
-}
-
-export enum GetArray {
-  BY_ID,
-  BY_NAME,
-}
-
-interface IPrimitiveCount {
-  points: number;
-  verts: number;
-  lines: number;
-  triangles: number;
-}
-
-interface IAbstractScalars {
-  cellFlag: boolean;
-}
-
-export interface IMapper2DInitialValues extends IAbstractMapperInitialValues {
-  arrayAccessMode?: number;
-  colorMode?: number;
+export interface IMapper2DInitialValues
+  extends IAbstractMapperInitialValues, IScalarColoringInitialValues {
   customShaderAttributes?: any;
   renderTime?: number;
-  scalarMode?: number;
-  scalarRange?: Range;
-  scalarVisibility?: boolean;
   static?: boolean;
 }
 
-export interface vtkMapper2D extends vtkAbstractMapper {
-  /**
-   * Create default lookup table. Generally used to create one when
-   * none is available with the scalar data.
-   */
-  createDefaultLookupTable(): void;
-
-  /**
-   *
-   * @param input
-   * @param {ScalarMode} scalarMode
-   * @param arrayAccessMode
-   * @param arrayId
-   * @param arrayName
-   */
-  getAbstractScalars(
-    input: any,
-    scalarMode: ScalarMode,
-    arrayAccessMode: number,
-    arrayId: any,
-    arrayName: any
-  ): IAbstractScalars;
-
-  /**
-   *
-   */
-  getArrayAccessMode(): number;
-
-  /**
-   * Get the array name to color by.
-   */
-  getColorByArrayName(): Nullable<string>;
-
-  /**
-   * Provide read access to the color array.
-   */
-  getColorMapColors(): Nullable<Uint8Array>;
-
-  /**
-   * Return the method of coloring scalar data.
-   */
-  getColorMode(): number;
-
-  /**
-   * Return the method of coloring scalar data.
-   */
-  getColorModeAsString(): string;
-
+export interface vtkMapper2D
+  extends
+    vtkAbstractMapper,
+    TScalarColoringWithout<
+      | 'clearColorArrays'
+      | 'getFieldDataTupleId'
+      | 'getInterpolateScalarsBeforeMapping'
+      | 'getNumberOfColorsInRange'
+      | 'setFieldDataTupleId'
+      | 'setInterpolateScalarsBeforeMapping'
+    > {
   /**
    *
    * @default []
@@ -100,42 +34,19 @@ export interface vtkMapper2D extends vtkAbstractMapper {
   getCustomShaderAttributes(): any;
 
   /**
-   * Get a lookup table for the mapper to use.
+   *
    */
-  getLookupTable(): any;
+  getPrimitiveCount(): IPrimitiveCount;
+
+  /**
+   *
+   */
+  getRenderTime(): number;
 
   /**
    * Get the transformCoordinate.
    */
   getTransformCoordinate(): any;
-
-  /**
-   * Return the method for obtaining scalar data.
-   */
-  getScalarMode(): number;
-
-  /**
-   * Return the method for obtaining scalar data.
-   */
-  getScalarModeAsString(): string;
-
-  /**
-   *
-   * @default [0, 1]
-   */
-  getScalarRange(): number[];
-
-  /**
-   *
-   * @default [0, 1]
-   */
-  getScalarRangeByReference(): number[];
-
-  /**
-   * Check whether scalar data is used to color objects.
-   * @default true
-   */
-  getScalarVisibility(): boolean;
 
   /**
    * Check whether the mapper’s data is static.
@@ -145,63 +56,9 @@ export interface vtkMapper2D extends vtkAbstractMapper {
 
   /**
    *
-   * @default false
-   */
-  getUseLookupTableScalarRange(): boolean;
-
-  /**
-   *
    * @default null
    */
   getViewSpecificProperties(): object;
-
-  /**
-   * Map the scalars (if there are any scalars and ScalarVisibility is on)
-   * through the lookup table, returning an unsigned char RGBA array. This is
-   * typically done as part of the rendering process. The alpha parameter
-   * allows the blending of the scalars with an additional alpha (typically
-   * which comes from a vtkActor, etc.)
-   * {
-   *     rgba: Uint8Array(),
-   *     location: 0/1/2, // Points/Cells/Fields
-   * }
-   * @param input
-   * @param {Number} alpha
-   */
-  mapScalars(input: any, alpha: number): void;
-
-  /**
-   *
-   * @param {Number} arrayAccessMode
-   */
-  setArrayAccessMode(arrayAccessMode: number): boolean;
-
-  /**
-   * Set the array name to color by.
-   * @param {String} colorByArrayName
-   */
-  setColorByArrayName(colorByArrayName: string): boolean;
-
-  /**
-   *
-   * @param {Number} colorMode
-   */
-  setColorMode(colorMode: number): boolean;
-
-  /**
-   * Sets colorMode to `DEFAULT`
-   */
-  setColorModeToDefault(): boolean;
-
-  /**
-   * Sets colorMode to `MAP_SCALARS`
-   */
-  setColorModeToMapScalars(): boolean;
-
-  /**
-   * Sets colorMode to `DIRECT_SCALARS`
-   */
-  setColorModeToDirectScalars(): boolean;
 
   /**
    * Sets point data array names that will be transferred to the VBO
@@ -210,107 +67,14 @@ export interface vtkMapper2D extends vtkAbstractMapper {
   setCustomShaderAttributes(customShaderAttributes: string[]): boolean;
 
   /**
-   * Set a lookup table for the mapper to use.
+   * Set the time required for the last render.
    */
-  setLookupTable(lookupTable: any): boolean;
+  setRenderTime(renderTime: number): boolean;
 
   /**
    * Set the transformCoordinate.
    */
   setTransformCoordinate(coordinate: any): boolean;
-
-  /**
-   * Control how the filter works with scalar point data and cell attribute
-   * data. By default (ScalarModeToDefault), the filter will use point data,
-   * and if no point data is available, then cell data is used. Alternatively
-   * you can explicitly set the filter to use point data
-   * (ScalarModeToUsePointData) or cell data (ScalarModeToUseCellData).
-   * You can also choose to get the scalars from an array in point field
-   * data (ScalarModeToUsePointFieldData) or cell field data
-   * (ScalarModeToUseCellFieldData). If scalars are coming from a field
-   * data array, you must call SelectColorArray before you call GetColors.
-   *
-   * When ScalarMode is set to use Field Data (ScalarModeToFieldData),
-   * you must call SelectColorArray to choose the field data array to
-   * be used to color cells. In this mode, the default behavior is to
-   * treat the field data tuples as being associated with cells. If
-   * the poly data contains triangle strips, the array is expected to
-   * contain the cell data for each mini-cell formed by any triangle
-   * strips in the poly data as opposed to treating them as a single
-   * tuple that applies to the entire strip. This mode can also be
-   * used to color the entire poly data by a single color obtained by
-   * mapping the tuple at a given index in the field data array
-   * through the color map. Use SetFieldDataTupleId() to specify
-   * the tuple index.
-   *
-   * @param scalarMode
-   */
-  setScalarMode(scalarMode: number): boolean;
-
-  /**
-   * Sets scalarMode to DEFAULT
-   */
-  setScalarModeToDefault(): boolean;
-
-  /**
-   * Sets scalarMode to USE_CELL_DATA
-   */
-  setScalarModeToUseCellData(): boolean;
-
-  /**
-   * Sets scalarMode to USE_CELL_FIELD_DATA
-   */
-  setScalarModeToUseCellFieldData(): boolean;
-
-  /**
-   * Sets scalarMode to USE_FIELD_DATA
-   */
-  setScalarModeToUseFieldData(): boolean;
-
-  /**
-   * Sets scalarMode to USE_POINT_DATA
-   */
-  setScalarModeToUsePointData(): boolean;
-
-  /**
-   * Sets scalarMode to USE_POINT_FIELD_DATA
-   */
-  setScalarModeToUsePointFieldData(): boolean;
-
-  /**
-   * Specify range in terms of scalar minimum and maximum (smin,smax). These
-   * values are used to map scalars into lookup table. Has no effect when
-   * UseLookupTableScalarRange is true.
-   *
-   * @param min
-   * @param max
-   * @default [0, 1]
-   */
-  setScalarRange(min: number, max: number): boolean;
-
-  /**
-   * Specify range in terms of scalar minimum and maximum (smin,smax). These
-   * values are used to map scalars into lookup table. Has no effect when
-   * UseLookupTableScalarRange is true.
-   *
-   * @param scalarRange
-   * @default [0, 1]
-   */
-  setScalarRange(scalarRange: number[]): boolean;
-
-  /**
-   *
-   * @param scalarRange
-   * @default [0, 1]
-   */
-  setScalarRangeFrom(scalarRange: number[]): boolean;
-
-  /**
-   * Turn on/off flag to control whether scalar data is used to color objects.
-   * @param {Boolean} scalarVisibility
-   * @default true
-   */
-  setScalarVisibility(scalarVisibility: boolean): boolean;
 
   /**
    * Turn on/off flag to control whether the mapper’s data is static. Static data
@@ -322,19 +86,6 @@ export interface vtkMapper2D extends vtkAbstractMapper {
    * @default false
    */
   setStatic(static: boolean): boolean;
-
-  /**
-   * Control whether the mapper sets the lookuptable range based on its
-   * own ScalarRange, or whether it will use the LookupTable ScalarRange
-   * regardless of it’s own setting. By default the Mapper is allowed to set
-   * the LookupTable range, but users who are sharing LookupTables between
-   * mappers/actors will probably wish to force the mapper to use the
-   * LookupTable unchanged.
-   *
-   * @param {Boolean} useLookupTableScalarRange
-   * @default false
-   */
-  setUseLookupTableScalarRange(useLookupTableScalarRange: boolean): boolean;
 
   /**
    * If you want to provide specific properties for rendering engines you can use
